Version 1.1.5
1. Bug when converting big decimal to int
2. Added tzs to currency list

+++ sdk/src/main/java/com/pesapal/sdk/fragment/auth/AuthFragment.kt
@@ -164,7 +164,7 @@ class AuthFragment : Fragment() {
 //                setErrorElements("Data -> Amount is missing")
                 setErrorElements("Error Code: 10201")
             }
-            else if(amount.toInt() < 1){
+            else if(amount.toBigDecimal() < BigDecimal(1)){
 //                setErrorElements("Data -> Amount is missing")
                 setErrorElements("Error Code: 102011")
             }
